+++ workflows/main.tf
@@ -17,7 +17,7 @@ resource "aws_lambda_layer_version" "lambda_dependencies" {
   layer_name       = "${local.prefix}-lambda_dependencies"
   source_code_hash = filebase64sha256("${var.DIST_DIR}/lambda_dependencies_layer.zip")
 
-  compatible_runtimes = ["python3.8"]
+  compatible_runtimes = [local.python_version]
 }
 
 resource "aws_lambda_function" "nop_lambda" {
@@ -29,5 +29,6 @@ resource "aws_lambda_function" "nop_lambda" {
   timeout          = 10
   source_code_hash = filebase64sha256("${var.DIST_DIR}/lambdas.zip")
 
-  runtime = "python3.8"
+  runtime = local.python_version
+  tags    = local.default_tags
 }