+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # CHANGELOG
 
 ## Unreleased
+* Update Lambda runtime to Python3.9
 * Tag resources using the aws provider level `default_tags` configuration
 
 ## v18.3.1.0